Make sierra_to_dynamo just be for bibs
