+++ build.sbt
@@ -37,8 +37,8 @@ lazy val id_minter = doSharedSetup(project, "catalogue_pipeline/id_minter")
 lazy val reindexer = doSharedSetup(project, "catalogue_pipeline/reindexer")
   .settings(libraryDependencies ++= Dependencies.reindexerDependencies)
 
-lazy val sierra_to_dynamo = doSharedSetup(project, "sierra_adapter/sierra_to_dynamo")
-  .settings(libraryDependencies ++= Dependencies.sierraToDynamoDependencies)
+lazy val sierra_bibs_to_dynamo = doSharedSetup(project, "sierra_adapter/sierra_bibs_to_dynamo")
+  .settings(libraryDependencies ++= Dependencies.sierraBibsToDynamoDependencies)
 
 lazy val sierra_bib_merger = doSharedSetup(project, "sierra_adapter/sierra_bib_merger")
   .settings(libraryDependencies ++= Dependencies.sierraBibMergerDepedencies)
@@ -54,7 +54,7 @@ lazy val root = (project in file("."))
     transformer,
     id_minter,
     reindexer,
-    sierra_to_dynamo,
+    sierra_bibs_to_dynamo,
     sierra_bib_merger,
     sierra_item_merger
   )
