+++ sierra_adapter/Makefile
@@ -4,15 +4,16 @@ SIERRA_ADAPTER = $(ROOT)/sierra_adapter
 include $(SIERRA_ADAPTER)/sierra_window_generator/Makefile
 include $(ROOT)/functions.Makefile
 
-sierra_to_dynamo-test:
-	$(call sbt_test,sierra_to_dynamo)
+sierra_bibs_to_dynamo-test:
+	$(call sbt_test,sierra_bibs_to_dynamo)
 
-sierra_to_dynamo-build: $(ROOT)/.docker/sbt_image_builder
-	TARGET=sierra_adapter/sierra_to_dynamo/target \
-	PROJECT=sierra_to_dynamo $(ROOT)/builds/run_sbt_image_build.sh
+sierra_bibs_to_dynamo-build: $(ROOT)/.docker/sbt_image_builder
+	TARGET=sierra_adapter/sierra_bibs_to_dynamo/target \
+	PROJECT=sierra_bibs_to_dynamo $(ROOT)/builds/run_sbt_image_build.sh
 
-sierra_to_dynamo-publish: sierra_to_dynamo-build
-	$(call publish_service,sierra_to_dynamo)
+sierra_bibs_to_dynamo-publish: sierra_bibs_to_dynamo-build
+	echo "Temporarily, does nothing (until ECR is in place)"
+	# $(call publish_service,sierra_bibs_to_dynamo)
 
 
 sierra_bib_merger-test:
@@ -39,13 +40,13 @@ sierra_item_merger-publish: sierra_item_merger-build
 
 sierra_adapter-test: \
 	sierra_window_generator-test \
-	sierra_to_dynamo-test \
+	sierra_bibs_to_dynamo-test \
 	sierra_bib_merger-test \
 	sierra_item_merger-test
 
 sierra_adapter-publish: \
 	sierra_window_generator-publish \
-	sierra_to_dynamo-publish \
+	sierra_bibs_to_dynamo-publish \
 	sierra_bib_merger-publish \
 	sierra_item_merger-publish
