+++ sierra_adapter/sierra_bibs_to_dynamo/docker-compose.yml
