+++ sierra_adapter/sierra_bibs_to_dynamo/src/main/resources/logback.xml
