+++ sierra_adapter/sierra_bibs_to_dynamo/src/main/scala/uk/ac/wellcome/platform/sierra_bibs_to_dynamo/modules/SierraDynamoConfigModule.scala
@@ -1,0 +1,15 @@
+package uk.ac.wellcome.platform.sierra_bibs_to_dynamo.modules
+
+import com.google.inject.Provides
+import uk.ac.wellcome.finatra.modules.DynamoConfigModule
+import uk.ac.wellcome.models.aws.DynamoConfig
+import javax.inject.Singleton
+
+object SierraDynamoConfigModule extends DynamoConfigModule {
+  val sierraToDynamoAppTable = flags("sierraToDynamo")
+
+  @Singleton
+  @Provides
+  def providesDynamoConfig(): DynamoConfig =
+    DynamoConfig(sierraToDynamoAppTable())
+}
