+++ sierra_adapter/sierra_bibs_to_dynamo/src/test/resources/logback-test.xml
