+++ sierra_adapter/sierra_bibs_to_dynamo/src/test/resources/wiremock/__files/items-21e1cd2e-d441-4fb4-acc8-2b2be7bcaad9.json
