+++ sierra_adapter/sierra_bibs_to_dynamo/src/test/resources/wiremock/__files/items-5da14461-d651-416f-b1f4-cc73e23a7165.json
