+++ sierra_adapter/sierra_bibs_to_dynamo/src/test/resources/wiremock/__files/items-6b3b4626-8c75-4426-a993-b905b79ef38a.json
