+++ sierra_adapter/sierra_bibs_to_dynamo/src/test/resources/wiremock/mappings/bibs-48b80ead-d492-40af-bace-97cd94cfbafd.json
