+++ sierra_adapter/sierra_bibs_to_dynamo/src/test/resources/wiremock/mappings/bibs-56afed67-47e7-4cf7-ad12-264ab5944e84.json
