+++ sierra_adapter/sierra_bibs_to_dynamo/src/test/resources/wiremock/mappings/items-3670d08b-9b74-40aa-b8d3-4a61f31ac993.json
