+++ sierra_adapter/sierra_bibs_to_dynamo/src/test/resources/wiremock/mappings/items-879008eb-cd0b-4a29-95ef-ea9b5b3d1620.json
